Configuração da base de testes de integração
