+++ teste/SME.SGP.TesteIntegracao/ConselhoDeClasseLancamento/Ao_inserir_nota_pos_conselho_bimestre.cs
@@ -1,16 +1,13 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Shouldly;
 using SME.SGP.Aplicacao;
 using SME.SGP.Dominio;
 using SME.SGP.Infra;
 using SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento.Base;
 using SME.SGP.TesteIntegracao.Setup;
-using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Xunit;
-using Shouldly;
 
 namespace SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento
 {
@@ -29,13 +26,12 @@ namespace SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento
             await CrieDados(
                 ObterPerfilProfessor(), 
                 COMPONENTE_CURRICULAR_PORTUGUES_ID_138, 
-                TipoNota.Nota, 
                 ANO_7, 
                 Modalidade.Fundamental, 
                 ModalidadeTipoCalendario.FundamentalMedio,
                 anoAnterior);
 
-            await ExecuteTeste(COMPONENTE_CURRICULAR_PORTUGUES_ID_138, anoAnterior);
+            await ExecuteTeste(ObtenhaDto(COMPONENTE_CURRICULAR_PORTUGUES_ID_138), anoAnterior, ALUNO_CODIGO_1, TipoNota.Nota);
         }
 
         [Theory]
@@ -46,13 +42,12 @@ namespace SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento
             await CrieDados(
                 ObterPerfilProfessor(), 
                 COMPONENTE_CURRICULAR_PORTUGUES_ID_138, 
-                TipoNota.Nota, 
                 ANO_7, 
                 Modalidade.Medio, 
                 ModalidadeTipoCalendario.FundamentalMedio,
                 anoAnterior);
 
-            await ExecuteTeste(COMPONENTE_CURRICULAR_PORTUGUES_ID_138, anoAnterior);
+            await ExecuteTeste(ObtenhaDto(COMPONENTE_CURRICULAR_PORTUGUES_ID_138), anoAnterior, ALUNO_CODIGO_1, TipoNota.Nota);
         }
 
         [Theory]
@@ -63,13 +58,12 @@ namespace SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento
             await CrieDados(
                 ObterPerfilProfessor(), 
                 COMPONENTE_CURRICULAR_PORTUGUES_ID_138, 
-                TipoNota.Nota, 
                 ANO_9, 
                 Modalidade.EJA, 
                 ModalidadeTipoCalendario.EJA,
                 anoAnterior);
 
-            await ExecuteTeste(COMPONENTE_CURRICULAR_PORTUGUES_ID_138, anoAnterior);
+            await ExecuteTeste(ObtenhaDto(COMPONENTE_CURRICULAR_PORTUGUES_ID_138), anoAnterior, ALUNO_CODIGO_1, TipoNota.Nota);
         }
 
         [Theory]
@@ -80,73 +74,48 @@ namespace SME.SGP.TesteIntegracao.ConselhoDeClasseLancamento
             await CrieDados(
                 ObterPerfilProfessor(), 
                 COMPONENTE_REGENCIA_CLASSE_FUND_I_5H_ID_1105, 
-                TipoNota.Nota, 
                 ANO_1, 
                 Modalidade.Fundamental, 
                 ModalidadeTipoCalendario.FundamentalMedio,
                 anoAnterior);
 
-            await ExecuteTeste(COMPONENTE_REGENCIA_CLASSE_FUND_I_5H_ID_1105, anoAnterior);
+            await ExecuteTeste(ObtenhaDto(COMPONENTE_REGENCIA_CLASSE_FUND_I_5H_ID_1105), anoAnterior, ALUNO_CODIGO_1, TipoNota.Nota);
         }
 
-        private async Task ExecuteTeste(long componente, bool anoAnterior)
+
+        private ConselhoClasseNotaDto ObtenhaDto(long componente)
         {
-            var comando = ServiceProvider.GetService<IComandosConselhoClasseNota>();
-            var dto = new ConselhoClasseNotaDto()
+            return new ConselhoClasseNotaDto()
             {
                 CodigoComponenteCurricular = componente,
                 Nota = 7,
                 Justificativa = JUSTIFICATIVA
             };
-            var dtoRetorno = await comando.SalvarAsync(dto, ALUNO_CODIGO_1, CONSELHO_CLASSE_ID, FECHAMENTO_TURMA_ID, TURMA_CODIGO_1, BIMESTRE_2);
-
-            dtoRetorno.ShouldNotBeNull();
-            var listaConselhoClasseNota = ObterTodos<ConselhoClasseNota>();
-            listaConselhoClasseNota.ShouldNotBeNull();
-            var classeNota = listaConselhoClasseNota.FirstOrDefault(nota => nota.ConselhoClasseAlunoId == 1);
-
-            if (anoAnterior)
-            {
-                classeNota.Nota.ShouldBe(7);
-            } else
-            {
-                classeNota.Nota.ShouldBeNull();
-                var listaAprovacaoNotaConselho = ObterTodos<WFAprovacaoNotaConselho>();
-                listaAprovacaoNotaConselho.ShouldNotBeNull();
-                var aprovacaoNotaConselho = listaAprovacaoNotaConselho.FirstOrDefault(aprovacao => aprovacao.ConselhoClasseNotaId == classeNota.Id);
-                aprovacaoNotaConselho.ShouldNotBeNull();
-                aprovacaoNotaConselho.Nota.ShouldBe(7);
-            }
-            
-            var listaConsolidado = ObterTodos<ConselhoClasseConsolidadoTurmaAluno>();
-            listaConsolidado.ShouldNotBeNull();
-            var consolidado = listaConsolidado.FirstOrDefault(consolidadoAluno => consolidadoAluno.AlunoCodigo == ALUNO_CODIGO_1);
         }
 
         private async Task CrieDados(
                         string perfil, 
                         long componente,
-                        TipoNota tipo, 
                         string anoTurma, 
                         Modalidade modalidade,
                         ModalidadeTipoCalendario modalidadeTipoCalendario,
                         bool anoAnterior)
         {
+            var dataAula = anoAnterior ? DATA_02_05_INICIO_BIMESTRE_2.AddYears(-1) : DATA_02_05_INICIO_BIMESTRE_2;
+
             var filtroNota = new FiltroNotasDto()
             {
                 Perfil = perfil,
-                Modalidade = modalidade ,
+                Modalidade = modalidade,
                 TipoCalendario = modalidadeTipoCalendario,
                 Bimestre = BIMESTRE_2,
                 ComponenteCurricular = componente.ToString(),
-                TipoNota = tipo,
                 AnoTurma = anoTurma,
-                ConsiderarAnoAnterior = anoAnterior
+                ConsiderarAnoAnterior = anoAnterior,
+                DataAula = dataAula
             };
-            var dataAula = anoAnterior ? DATA_02_05_INICIO_BIMESTRE_2.AddYears(-1) : DATA_02_05_INICIO_BIMESTRE_2;
-
+            
             await CriarDadosBase(filtroNota);
-            await CriarAula(filtroNota.ComponenteCurricular, dataAula, RecorrenciaAula.AulaUnica, NUMERO_AULA_1);
         }
     }
 }
